Release v2.0.0

+++ src/main/java/si/deisinger/providers/model/ampeco/AmpecoLocationPins.java
@@ -3,10 +3,10 @@ package si.deisinger.providers.model.ampeco;
 import java.util.ArrayList;
 
 public class AmpecoLocationPins {
-	public ArrayList<Pin> pins;
+    public ArrayList<Pin> pins;
 
-	public static class Pin {
-		public int id;
-		public String geo;
-	}
+    public static class Pin {
+        public Long id;
+        public String geo;
+    }
 }
